1.fix existed times count bug 2.change tracing hugepages from cpu_cache to filler 3.test redis and phoronix 4.set malloc_release_bytes_per_sec 0.5MB

+++ tcmalloc/cpu_cache.cc
@@ -614,7 +614,7 @@ uint32_t CPUCache::PerClassResizeInfo::Tick() {
   return state.quiescent_ticks - 1;
 }
 
-CpuStats CPUCache::GetCpuStats(std::map<uint64_t,uint64_t> &hpMap) {
+CpuStats CPUCache::GetCpuStats(std::map<uintptr_t,size_t> &hpMap) {
   return freelist_.GetSlabStats(hpMap);
 }
